// src/types/Enum.ts
export enum FiltersEnum {
  ALL = "ALL",
  FINISH = "FINISH",
  ACTIVE = "ACTIVE"
}

export enum LevelEnum {
  HIGHEST = "HIGHEST",
  IMPORTANT = "IMPORTANT",
  ADVANCED = "ADVANCED",
  NORMAL = "NORMAL"
}

export enum CyclicEnum {
  NONE = "NONE",
  ONEDAY = "ONEDAY",
  ONEWEEK = "ONEWEEK",
  ONEMONTH = "ONEMONTH",
  ONEQUARTER = "ONEQUARTER",
  ONEYEAR = "ONEYEAR"
}


export enum langEnum {
  zh = "zh",
  en = "en"
}